feat: remove warning (#104)
Remove a warning that a categorical column with many categories is used
for coloring of TS line plot.
Following #94, which removes most of the warnings, the warning here is
removed since it does not bring relevant information to the user (the
user can see that there are many colors), and makes the final reports
ugly.