feat!: filter and check columns in section main functions
Previously columns were computed in `__init__`s of the respective subsections of multivariate analysis.
This required that the DataFrame, on which the analysis is performed, is specified in the constructor,
which prevented creating the section without the DataFrame.
This change makes https://github.com/datamole-ai/edvart/issues/36 much easier and enables removal
of the `add_<section>` methods.