feat!: filter and check columns in section main functions (#133)
Previously columns were computed in `__init__`s of the respective
subsections of multivariate analysis.
This required that the DataFrame, on which the analysis is performed, is
specified in the constructor, which prevented creating the section
without the DataFrame.
This change makes https://github.com/datamole-ai/edvart/issues/36 much
easier and enables removal of the `add_<section>` methods.

BREAKING CHANGE: parameter `df` removed from constructors of:
- `edvart.report_sections.multivariate_analysis.MultivariateAnalysis`,
- `edvart.report_sections.multivariate_analysis.PCA`,
- `edvart.report_sections.multivariate_analysis.ParallelCoordinates`,
- `edvart.report_sections.multivariate_analysis.ParallelCategories`,
- `edvart.report_sections.umap.UMAP`.

BREAKING CHANGE: parameter `nunique_max` removed from constructors of
`edvart.multivariate_analysis.ParallelCategories` and
`edvart.multivariate_analysis.ParallelCoordinates`. The value is now
hard-coded. Specify columns manually to override the default choice of
columns, which was influenced by the parameter `nunique_max`.